requirements.txt
python-telegram-bot==21.3
fastapi==0.110.2
uvicorn[standard]==0.29.0
pydraughts==0.6.7
google-generativeai==0.3.2
python-dotenv==1.0.1
pyngrok==6.0.0
requests==2.32.0